% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_laspeyeres_index.R
\name{get_laspeyeres_index}
\alias{get_laspeyeres_index}
\title{get_laspeyeres_index Calculates the Laspeyeres index for each commune with 2010 as the reference year}
\usage{
get_laspeyeres_index(dataset, start_year = "2010")
}
\arguments{
\item{dataset}{Flat data df as returned by clean_flat_data()}

\item{start_year}{Optional: Minimal year to consider. Defaults to 2009 because price data starts in 2010}
}
\value{
The input dataset with added columns: p0, p0_m2, pl and pl_m2. p0 are prices at year 0 (2010), and pl are the index.
}
\description{
get_laspeyeres_index Calculates the Laspeyeres index for each commune with 2010 as the reference year
}
\examples{
\dontrun{
commune_level_data_laspeyeres <- get_laspeyeres(commune_level_data)
} 
}
