% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAP.BrunoR.nosync-package.R
\docType{package}
\name{RAP.BrunoR.nosync-package}
\alias{RAP.BrunoR.nosync}
\alias{RAP.BrunoR.nosync-package}
\title{RAP.BrunoR.nosync: Housing Data For Luxembourg}
\description{
This package contains functions to get, clean and analyse housing price data for Luxembourg.
}
\author{
\strong{Maintainer}: Jake Tufts \email{jakestufts@aol.co.uk} (\href{https://orcid.org/0009-0008-3798-8805}{ORCID})

}
\keyword{internal}
